print binary

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -8,8 +8,8 @@ int _putchar(char c);
 void print_binary(unsigned long int n)
 {
 	int i;
-	int numBits = sizeof(unsigned int) * 8;
-	unsigned int mask;
+	int numBits = sizeof(unsigned long int) * 8;
+	unsigned long int mask;
 	int printStarted = 0;
 
 	if (n == 0)
